Motivo-backend/motivoproject/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
gunicorn==21.2.0
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
